pc - remove line from init.sh that deletes init.sh after running
It doesn't make sense to get rid of this file after running once,
since the operation needs to be done again any time you change from one
commit to another (e.g. when doing 'git bisect' to troubleshoot a problem.

In addition, it creates a 'change' that has to be 'undone' if you are
committing something to GitHub.  This is confusing for novices, which
is at odds with the stated goal of allolib_playground, i.e. to be
a place where beginners can get started with as little confusion as possible.